Create PracticeSessions
The racers must practice for the race. Your job is to keep the records of the roads and the time for each lap. The track with the best time will be the chosen one for the finals.
 
Write a program, that keeps information about roads and the racers who practice on them.  When the practice begins, you’re going to start receiving data until you get the "END" message. There are three possible commands:
•	"Add->{road}->{racer}"
o	Add the road if it doesn't exist in your collection and add the racer to it.
•	"Move->{currentRoad}->{racer}->{nextRoad}"
o	Find the racer on the current road and move him to the next one, only if he exists in the current road. Both roads will always be valid and will already exist.
•	"Close->{road}"
o	Find the road and remove it from the sessions, along with the racers on it if it exists.
In the end, print all of the roads with the racers who have practiced and ordered by the count of the racers in descending order, then by the roads in ascending order. The output must be in the following format:
Practice sessions:
{road}
++{racer}
++{racer}
++{racer}
………………………..
Input / Constraints
•	You will be receiving lines of information in the format described above, until you receive the "END" command.
•	The input will always be in the right format.
•	Both roads from the "Move" command will always be valid and you don't need to check them explicitly. 
Output
•	Print the roads with their racers in the format described above.